% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_clusters.R
\name{ckmean_getlabel}
\alias{ckmean_getlabel}
\title{This function uses kmeans to find the cluster labels}
\usage{
ckmean_getlabel(x, k = 5, att_type = "numeric")
}
\arguments{
\item{x}{numeric vector}

\item{k}{maximum number of clusters}
}
\value{
cluster labels
}
\description{
This function uses kmeans to find the cluster labels
}
