% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_compression.R
\name{cluster_rules}
\alias{cluster_rules}
\title{Ensemble Rule Compression}
\usage{
cluster_rules(rules, k)
}
\arguments{
\item{rules}{character vector containing the decision rules}

\item{k}{maximum number of clusters}
}
\value{
list with conditions and rules
}
\description{
Ensemble Rule Compression
}
