% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cre.R
\name{cre}
\alias{cre}
\title{Learn a compressed rule ensemble}
\usage{
cre(
  x,
  y,
  k = 4,
  treetype = "XGB",
  standardize = F,
  linear = T,
  alpha = 1,
  task = "regression",
  min_sup = 0.05,
  forest_control = NULL,
  model_type = "glmnet",
  eta = 0.5
)
}
\arguments{
\item{x}{input matrix}

\item{y}{outcome vector}

\item{treetype}{foresttype to generate the rules. Currently XGBoost and RandomForest are supported}

\item{standardize}{Should the rules be standardized? This essentiall removes the extra penalty on low support/high complexity rules. Can be sometimes beneficial.}

\item{linear}{Boolean if linear terms should be included}

\item{alpha}{alpha parameter in glmnet, alpha = 1 for lasso alpha = 0 for ridge.}

\item{task}{"regression" or "class" for (binary) classification}

\item{forest_control}{parameters that are passed to xgboost}
}
\value{
list containing the cre model
}
\description{
Learn a compressed rule ensemble
}
