% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_duplicates.R
\name{delete_duplicates}
\alias{delete_duplicates}
\title{Removes colinear and redundant rules. In case of colinearity (r = 1) the rule with higher frequency in the forest is kept}
\usage{
delete_duplicates(x, rules_frame)
}
\arguments{
\item{x}{rule matrix}
}
\value{
rules to be removed
}
\description{
Removes colinear and redundant rules. In case of colinearity (r = 1) the rule with higher frequency in the forest is kept
}
