% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genrulesXGB.R
\name{genrulesXGB}
\alias{genrulesXGB}
\title{Extracts rules from a XGB model.}
\usage{
genrulesXGB(x, y, forest_control = NULL, task = "regression")
}
\arguments{
\item{x}{input matrix}

\item{y}{outcome vector}

\item{forest_control}{parameters that are passed to xgboost}

\item{task}{"regression" or "class" for (binary) classification}
}
\value{
list of rules extracted from the xgboost model
}
\description{
Extracts rules from a XGB model.
}
