% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/important_rules.R
\name{important_rules}
\alias{important_rules}
\title{Prints the most important rules found be the CRE model}
\usage{
important_rules(model, s = "lambda.1se")
}
\arguments{
\item{model}{cre model}

\item{s}{lambda.1se or lambda.min}
}
\value{
prints the most important rules
}
\description{
Prints the most important rules found be the CRE model
}
