% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_X.R
\name{transformX}
\alias{transformX}
\title{Transforms the input matrix to the Rule Matrix, using the ensemble rule approach.}
\usage{
transformX(x, rules_frame)
}
\arguments{
\item{x}{input matrix}

\item{rules_frame}{data frame containing the ensemble rules as outputted by ensemble_compression}
}
\value{
design matrix with 1 column per rule
}
\description{
Transforms the input matrix to the Rule Matrix, using the ensemble rule approach.
}
