% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_clusters.R
\name{visualise_clusters}
\alias{visualise_clusters}
\title{Plots the distribution of splitpoints in each covariate}
\usage{
visualise_clusters(model, n_most = 6)
}
\arguments{
\item{model}{cre model}

\item{n_most}{How many covariates to show. Takes the n_most covariates with the most splitpoints.}
}
\value{
histogram of splitpoint distribution
}
\description{
Plots the distribution of splitpoints in each covariate
}
